Remove unused files.
Reorganized remaining files.
