--- src/components/cards/PrereqCard.svelte
@@ -1,13 +1,0 @@
-<script>
-    export let title;
-    export let content;
-  </script>
-  
-  <div class="ps-10 pt-5">
-    <h3 class="text-4xl font-extrabold text-red-800 text-center">{title}:</h3>
-    <div class="max-w-3xl mx-auto mt-5 text-center">
-      <p class="text-lg text-slate-600 text-slate-400">
-        {content}
-      </p>
-    </div>
-  </div>
