--- src/components/dropdown/Vertical.svelte
@@ -1,31 +1,0 @@
-<script>
-  import Icon from '@iconify/svelte'
-  import LightButton from '../buttons/LightButton.svelte'
-  export let dropdown
-  export let title
-  export let href
-</script>
-
-<div>
-  <div class="relative inline-block dropdown">
-    <LightButton icon="mdi:chevron-down" text={title} {href} />
-    <ul
-      class="absolute pt-2 hidden text-gray-700 rounded w-80 dropdown-menu"
-    >
-      {#each dropdown as item}
-        <li>
-          <a
-            class="block text-slate-500 bg-white hover:text-black bg-black hover:bg-battle-red hover:text-white font-bold px-4 py-2 whitespace-no-wrap hover:bg-white hover:outline hover:rounded hover:outline-red-800"
-            href={item.href}>{item.text}</a
-          >
-        </li>
-      {/each}
-    </ul>
-  </div>
-</div>
-
-<style>
-  .dropdown:hover .dropdown-menu {
-    display: block;
-  }
-</style>
