--- src/components/elements/PageHero.astro
@@ -1,32 +1,0 @@
----
-import Hero from './Hero.svelte'
-
-interface Props {
-  title: string
-  content?: string
-  link?: string
-  linkText?: string
-  padding?: string
-  location?: string
-}
-const { link, padding, content, title, location = 'left' } = Astro.props
----
-
-<div class="flex h-screen">
-  <!-- Left Column -->
-  <div class="w-1/2 p-4">
-    <Hero
-      title={title}
-      content={content}
-      link={link}
-      padding={padding}
-      linkText={location}
-    />
-  </div>
-
-  <!-- Right Column -->
-  <div
-    class="w-1/2 p-4 bg-cover bg-no-repeat bg-slate-200 bg-[url('https://dummyimage.com/1080/fff/aaa')]"
-  >
-  </div>
-</div>
