--- src/components/sections/Vision.svelte
@@ -1,36 +1,0 @@
-<script>
-  import Hero from '../elements/Hero.svelte'
-  import ButtonCustom from '../buttons/ButtonCustom.svelte'
-</script>
-
-<section class="">
-  <div
-    class="gap-16 items-center py-8 px-4 mx-auto max-w-screen-xl lg:grid lg:grid-cols-2 lg:py-16 lg:px-6">
-    <div class="font-light text-gray-500 sm:text-lg text-gray-400">
-      <h2
-        class="mb-4 text-4xl tracking-tight font-extrabold text-gray-900 text-white text-center">
-        WHAT WE DO
-      </h2>
-      <p class="mb-4">
-        Lorem ipsum dolor sit amet. Et quae quibusdam sed consequatur iste nam
-        asperiores facilis qui error iste qui quia tenetur est rerum dolorem.
-        Eum eveniet placeat cum impedit officiis est odit asperiores. Lorem
-        ipsum dolor sit amet. Et quae quibusdam sed consequatur iste nam
-        asperiores facilis qui error iste qui quia tenetur est rerum dolorem.
-        Eum eveniet placeat cum impedit officiis est odit asperiores.
-      </p>
-      <div class="grid w-full gap-3 mt-8 sm:inline-flex sm:justify-center">
-        <ButtonCustom color="white" size="xl" customClasses="w-64" link="#">
-          LEARN MORE
-        </ButtonCustom>
-      </div>
-    </div>
-
-    <div class="grid grid-cols-1 gap-4 mt-8">
-      <img
-        class="w-full rounded-lg"
-        src="https://dummyimage.com/1080/fff/aaa"
-        alt="office content 1" />
-    </div>
-  </div>
-</section>
