--- src/components/utils/Padding.astro
@@ -1,9 +1,0 @@
----
-interface Props {
-  classes: string
-}
-
-const { classes } = Astro.props
----
-
-<div class:list={["", classes]}></div>
