+++ src/content/aboutMetrics/missionMetrics.yaml
@@ -1,14 +1,14 @@
 metrics:
   - metric: 'Marines Trained'
-    icon: '/icons/icon homepage marines trained.svg'
+    icon: '/icons/marinesTrained.svg'
     value: '4'
   - metric: "Applications Deployed"
-    icon: '/icons/icon homepage applications deployed.svg'
+    icon: '/icons/applicationsDeployed.svg'
     value: '4'
   - metric: 'Units Using MCSWF Applications'
-    icon: '/icons/icon homepage units using.svg'
+    icon: '/icons/unitsUsing.svg'
     value: '16'
   - metric: 'Average Days 0-1 MVP'
-    icon: '/icons/icon homepage average days.svg'
+    icon: '/icons/averageDays.svg'
     value: '81'
