+++ src/content/products/my-career.yaml
@@ -6,9 +6,6 @@ background:
 image: 
   src: '/images/myCareer.png'
   alt: ' '
-secondImage:
-  src: ''
-  alt: ''
 order: 2
 heading: "Manpower Management Enlisted Assignments"
 bullets:
