--- src/content/roleDetails/prod_designer.md
@@ -1,23 +1,0 @@
----
-title: 'ABOUT THE JOB'
-role: PRODUCT DESIGNER
-url: product_designer
-about: Et quae quibusdam sed consequatur iste nam asperiores facilis qui error iste qui quia tenetur est rerum dolorem. Eum eveniet placeat
----
-# ABOUT THE JOB
-  Lorem ipsum dolor sit amet. Et quae quibusdam sed consequatur iste nam asperiores facilis qui error iste qui quia tenetur est rerum dolorem. Eum eveniet placeat cum impedit officiis est odit asperiores.  Lorem ipsum dolor sit amet. Et quae quibusdam sed consequatur iste nam asperiores facilis qui error iste qui quia tenetur est rerum dolorem. Eum eveniet placeat cum impedit officiis est odit asperiores.
-  - Lorem ipsum dolor sit amet. 
-  - Et quae quibusdam sed consequatur iste nam asperiores facilis qui error iste qui quia tenetur est rerum dolorem. Eum eveniet placeat
-  - cum impedit officiis est odit asperiores.  Lorem ipsum dolor sit amet. Et quae quibusdam sed consequatur iste nam asperiores facilis 
-  - qui error iste qui quia tenetur est rerum dolorem. Eum eveniet placeat cum impedit officiis est odit asperiores.
-# PLATFOROM ENGINEER RESPONSIBILITIES
-  - Lorem ipsum dolor sit amet. 
-  - Et quae quibusdam sed consequatur iste nam asperiores facilis qui error iste qui quia tenetur est rerum dolorem. Eum eveniet placeat
-  - cum impedit officiis est odit asperiores.  Lorem ipsum dolor sit amet. Et quae quibusdam sed consequatur iste nam asperiores facilis 
-  - qui error iste qui quia tenetur est rerum dolorem. Eum eveniet placeat cum impedit officiis est odit asperiores.
-# WHATS IN IT FOR ME?
-  - Lorem ipsum dolor sit amet. 
-  - Et quae quibusdam sed consequatur iste nam asperiores facilis qui error iste qui quia tenetur est rerum dolorem. Eum eveniet placeat
-  - cum impedit officiis est odit asperiores.  Lorem ipsum dolor sit amet. Et quae quibusdam sed consequatur iste nam asperiores facilis 
-  - qui error iste qui quia tenetur est rerum dolorem. Eum eveniet placeat cum impedit officiis est odit asperiores.
----
