+++ src/pages/about.astro
@@ -2,7 +2,7 @@
 import Layout from '../layouts/Layout.astro';
 import Hero from '../components/elements/Hero.svelte';
 import ContentImageFull from '../components/sections/ContentImageFull.svelte';
-import FooterFull from '../components/footer/FooterFull.svelte';
+import FooterFull from '@components/FooterFull.svelte';
 import ButtonCustom from '../components/buttons/ButtonCustom.svelte';
 import MetricCard from '../components/cards/MetricCard.svelte';
 import { getCollection, type CollectionEntry } from 'astro:content';
