+++ src/pages/career.astro
@@ -6,7 +6,7 @@ import MetricCard from '../components/cards/MetricCard.svelte';
 
 import '../styles/global.css'
 import '../styles/fonts.css'
-import FooterFull from '../components/footer/FooterFull.svelte'
+import FooterFull from '@components/FooterFull.svelte'
 import ContentImageFull from '../components/sections/ContentImageFull.svelte'
 import CareerAccordion from '../components/accordion/CareerAccordion.svelte'
 import ButtonCustom from '@components/buttons/ButtonCustom.svelte'
@@ -72,7 +72,7 @@ displaySiteBanner.set(false)
             content="Every member of the team is essential to the rapid development of an application that satisfies the critical requirements of the Fleet Marine Forces. Teams remain together throughout the software development lifecycle. For approximately nine months, they are paired with an industry expert for daily tasks, ensuring that they are using the technology industry's best practices."
             link=""
             linkText=""
-            image="/images/6414552 1.png"
+            image="/images/6414552_1.png"
             imageAltText="as"
           />
         </div>
