+++ src/pages/cohort.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '@layouts/Layout.astro';
-import CohortGroup from '@components/cards/CohortGroup.svelte';
+import CohortGroup from '@components/sections/CohortGroup.svelte';
 import '../styles/global.css';
 
 import { getCollection } from 'astro:content';
