+++ src/pages/index.astro
@@ -5,7 +5,7 @@ import Countdown from '../components/elements/Countdown.svelte';
 import Products from '../components/sections/Products.svelte';
 import '../styles/global.css';
 import '../styles/fonts.css';
-import FooterFull from '../components/footer/FooterFull.svelte';
+import FooterFull from '@components/FooterFull.svelte';
 import ContentImageFull from '../components/sections/ContentImageFull.svelte';
 import { getCollection } from 'astro:content'; // Get products collection
 import { displaySiteBanner } from '../components/stores.js';
