+++ src/pages/products/[...product].astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '../../layouts/Layout.astro';
 import Hero from '../../components/elements/Hero.svelte';
-import FooterFull from '../../components/footer/FooterFull.svelte';
+import FooterFull from '@components/FooterFull.svelte';
 import MetricCard from '../../components/cards/MetricCard.svelte';
 import { getCollection, type CollectionEntry } from 'astro:content';
 import { displaySiteBanner } from '../../components/stores.js';
