+++ src/pages/roles/[...role].astro
@@ -3,8 +3,8 @@ import Layout from '../../layouts/Layout.astro';
 import Hero from '../../components/elements/Hero.svelte';
 
 import '../../styles/global.css';
-import FooterFull from '../../components/footer/FooterFull.svelte';
-import CohortGroup from '../../components/cards/CohortGroup.svelte';
+import FooterFull from '@components/FooterFull.svelte';
+import CohortGroup from '../../components/sections/CohortGroup.svelte';
 import { getCollection } from 'astro:content';
 import { displaySiteBanner } from '../../components/stores.js';
 import { externalLinks } from '@content/constants';
@@ -44,8 +44,8 @@ const entries = await getCollection('cohorts');
     </div>
     <div class="relative z-20 max-h-screen scrollbar-hide">
       <CohortGroup entries={entries} client:only />
-      <div class="">
-        <FooterFull client:only background="bg-[#f0f0f0]" />
+      <div class="relative z-20">
+        <FooterFull bg-white client:only/>
       </div>
     </div>
   </div>
