+++ src/pages/roles/index.astro
@@ -4,7 +4,7 @@ import Hero from '@components/elements/Hero.svelte';
 
 import '../../styles/global.css';
 import '../../styles/fonts.css';
-import FooterFull from '../../components/footer/FooterFull.svelte';
+import FooterFull from '@components/FooterFull.svelte';
 import ContentImageFull from '../../components/sections/ContentImageFull.svelte';
 const posts = await Astro.glob('../../content/roles/*.md');
 import { displaySiteBanner } from '../../components/stores.js';
